services/orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatif-orchestrator"
version = "0.1.0"
description = "WhatIf orchestrator with Vertex-backed agent runtime"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "orjson>=3.10.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "google-genai>=1.65.0",
  "google-adk>=1.18.0",
  "upstash-redis>=1.5.0",
  "google-cloud-firestore>=2.23.0",
  "google-cloud-pubsub>=2.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
  "integration: tests that call live external services",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["tests"]

[[tool.mypy.overrides]]
module = ["google.cloud", "google.cloud.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["upstash_redis", "upstash_redis.*"]
ignore_missing_imports = true
